refactor side menu component and update import path
