+++ src/app/(aplicacao)/criar/page.tsx
@@ -1,0 +1,7 @@
+export default function Home() {
+  return (
+    <>
+      <div className=" bg-corNeutro">Estou em criar</div>
+    </>
+  );
+}
