+++ src/app/(aplicacao)/layout.tsx
@@ -1,25 +1,13 @@
-'use client';
-import { useCallback, useState } from 'react';
-import SideMenu from '@/components/sideMenu/SideMenu';
+import MenuContainer from '../../components/sideMenu/menuContainer';
 
 export default function LayoutAuthenticated({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  const [isOpen, setIsOpen] = useState<boolean>(false);
-
-  const toggleSidebar = useCallback(() => {
-    setIsOpen((prev) => !prev);
-  }, []);
-
   return (
     <div className="flex h-screen bg-corNeutro">
-      <SideMenu
-        isOpen={isOpen}
-        toggleSidebar={toggleSidebar}
-      />
-      <div className={`flex-grow transition-all ${isOpen ? 'ml-64' : 'ml-20'}`}>{children}</div>
+      <MenuContainer>{children}</MenuContainer>
     </div>
   );
 }
